requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
tiktoken>=0.5.1
numpy>=1.24.0
tqdm>=4.66.0
pytest>=7.4.0

